+++ SDKs/iOS/Debug/EOCore.xcframework/ios-arm64_armv7/EOCore.framework/Headers/EOMessage.h
@@ -2,7 +2,7 @@
 //  EOMessage.h
 //  EOCore
 //
-// Copyright (C) 2021 Acoustic, L.P. All rights reserved.
+// Copyright (C) 2022 Acoustic, L.P. All rights reserved.
 //
 // NOTICE: This file contains material that is confidential and proprietary to
 // Acoustic, L.P. and/or other developers. No license is granted under any intellectual or
@@ -21,7 +21,7 @@
  * @discussion EOCore will update this offset automatically when the message is posted to the queue.
  * @return The offset in seconds.
  */
-@property (nonatomic) unsigned long long offset;
+@property (nonatomic) double offset;
 
 /*!
  * @abstract The logging level of the EOMessage.
